requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
flask-caching==2.3.0
plotly==5.24.1
pandas==2.2.2
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.2
